requirements.txt
fastapi
uvicorn
jinja2
pandas
numpy
scikit-learn
pymongo
motor
dnspython
python-dotenv
openai